% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{classified_plots}
\alias{classified_plots}
\title{Plot the results}
\usage{
classified_plots(
  coh_df,
  classification_list = c(0.2, 0.4, 0.6, 0.8, 1),
  number_of_images,
  project_path
)
}
\arguments{
\item{coh_df}{DataFrame containing the results in the last columns.  It is expected to have as many result columns as images to analyse.}

\item{classification_list}{A vector of five entries defining the classification boundaries. Set the lowest boundary at first. Default is c(0.2, 0.4, 0.6, 0.8, 1).}

\item{number_of_images}{A numerical value representing the number of input images.}

\item{project_path}{Path for saved results.}
}
\value{
//
}
\description{
This function takes a DataFrame containing all calculated results and removes all columns except the results. All results are classified with a given classification scheme.
After that, a bar plot is created, showing the structure of classified buildings in relation to 100 \%. A line plot is created to show the absolute numbers of classified buildings.
Both plots plot the development over time as every column is plotted. The plots are saved as .png in one image.
}
\examples{
image_count <- clips[[3]]
final_dir <- '[...]/GitHub/package_test/clipped_data/'

classified_plots(coh_df = coh_results, number_of_images = image_count, project_path = final_dir)

}
