% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{coh_calc}
\alias{coh_calc}
\title{Calculate the avergae coherence value per buildings polygon.}
\usage{
coh_calc(rast_data, buildings, target_crs = "EPSG:4326", project_path)
}
\arguments{
\item{rast_data}{List of SpatRaster data.}

\item{buildings}{Dataframe containing building geometries.}

\item{target_crs}{User defined CRS as 'EPSG:123456'. Default is EPSG:4326.}

\item{project_path}{Path for saved results.}
}
\value{
Single DataFrame containing calculated coherence values appended to the original DF.
}
\description{
This function overlays given raster objects with the building polygons and calculates an average coherence value per polygon.
Results are added to the building DF and saved as .gpkg.
}
\examples{
gaza_crs <- 'EPSG:32636' 
final_dir <- '[...]/GitHub/package_test/clipped_data/'

clipped_raster <- clips[[1]]
clipped_buildings <- clips[[2]]

coh_results <- coh_calc(rast_data = clipped_raster, buildings = clipped_buildings, target_crs = gaza_crs, project_path = final_dir)

}
