% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coherence.R
\name{coherence}
\alias{coherence}
\title{Magnitude Square Coherence Estimation}
\usage{
coherence(x, window = 300, overlap = 0.5, samplingFreq = 1000, nfft = 2000)
}
\arguments{
\item{x}{A data frame with two numeric columns representing the two signals to analyze.}

\item{window}{Integer specifying the length of the Hann window for segmenting the signals. Default is 300.}

\item{overlap}{Numeric value specifying the number of overlapping samples between adjacent segments.
Should be a fraction of the window length, typically between 0 and `window`. Default is calculated as `cohOverlap * length(hann(window)) / samplingFreq`.}

\item{samplingFreq}{Sampling frequency of the signals in Hz (samples per second). Default is 1000.}

\item{nfft}{Integer specifying the number of points used in the FFT computation. Should be a power of 2 and at least as large as the window length. Default is 2000.}
}
\value{
A list containing:
  \item{f}{Frequency vector}
  \item{Cxy}{Coherence estimate between the two signals}
}
\description{
Computes the magnitude square coherence between two signals using the `gsignal::mscohere` function.
This function estimates the coherence between two signals based on the Hann window and given parameters.
}
