% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coherence.R
\name{fisherz}
\alias{fisherz}
\title{Fisher z-transformation}
\usage{
fisherz(x)
}
\arguments{
\item{x}{Numeric vector or matrix containing Pearson correlation coefficients (r) to be transformed.}
}
\value{
A numeric vector or matrix with transformed values.
}
\description{
Fisher's z-transformation converts Pearson's r to the normally distributed variable z.
The formula for the transformation is:
\deqn{z_r = \tanh^{-1}(r) = \frac{1}{2} \log \left( \frac{1+r}{1-r} \right)}
}
