% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{load_and_clip}
\alias{load_and_clip}
\title{Load and clip .tif files with building polygons.}
\usage{
load_and_clip(
  data_path,
  target_crs = "EPSG:4326",
  buildings_path,
  save_clips = FALSE,
  project_path
)
}
\arguments{
\item{data_path}{Path to the raw data containing coherence maps in .tif format.}

\item{target_crs}{User defined CRS as 'EPSG:123456'. Default is EPSG:4326.}

\item{buildings_path}{Path to the raw builing vector file.}

\item{save_clips}{Set to TRUE or FALSE, if the results should be saved. Default is TRUE.}

\item{project_path}{Path for saved results.}
}
\value{
List of SpatRaster objects

DataFrame with all clipped buildings

Single numerical value representing the number of images
}
\description{
This function takes .tif files and converts them to a specified CRS defined by the user, if necessary. The function checks for an extent as well to exclude faulty
images. Buiding polygons are read in and transformed to the same CRS. The raster's extent is used to clip the extent od the buildings. After that, the polygons are used to clip
the raster itself. The function stores and returns the clipped .tif files, the clipped buildings and returns the number of images loaded.
}
\examples{
raw_path <- '[...]/GitHub/package_test/raw_data/'
gaza_crs <- 'EPSG:32636' 
final_dir <- '[...]/GitHub/package_test/clipped_data/'
gaza_buildings <- '[...]/GitHub/package_test/raw_data/Gaza_Stripe_buildings.shp'

clips <- load_and_clip(data_path = path, target_crs = gaza_crs, buildings_path = gaza_buildings, save_clips = TRUE, project_path = final_dir)

 
}
